+++ metadata.rb
@@ -1,7 +1,9 @@
 name             'snowflake-nativex'
-maintainer       'simplyadrian'
-maintainer_email 'simplyadrian@gmail.com'
+Maintainer       'NativeX'
+maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
 description      'Installs/Configures snowflake-nativex'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
+
+depends 'git'